+++ types/Object/FullObject.d.ts
@@ -1,7 +1,0 @@
-/**
- * Marker class used by `to(FullObject)` conversion.
- *
- * @class FullObject
- */
-export default class FullObject {
-}