+++ types/Object/UndefinedObject.d.ts
@@ -1,7 +1,0 @@
-/**
- * Simple placeholder class used to retain undefined values during conversion.
- *
- * @class UndefinedObject
- */
-export default class UndefinedObject {
-}