+++ types/Parser/index.d.ts
@@ -1,4 +1,0 @@
-export default Parser;
-import Node from './Node.js';
-import Parser from './Parser.js';
-export { Node, Parser };